producer/src/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
confluent-kafka==2.4.0
fastavro==1.9.5
gevent==24.2.1
greenlet==3.0.3
idna==3.7
requests==2.32.3
urllib3==2.2.2
websocket-client==1.8.0
zope.event==5.0
zope.interface==6.4.post2
